refactor(shared): centralize row map helpers
